refactor: Transform ContractingDates record to static FakeContractDates class
