+++ Chapter-4-applying-tactical-domain-driven-design/Fitnet.Contracts/Src/Fitnet.Contracts.Core.UnitTests/PrepareContract/Builders/ContractBuilder.cs
@@ -1,10 +1,13 @@
 ﻿namespace EvolutionaryArchitecture.Fitnet.Contracts.Core.UnitTests.PrepareContract;
 
+using Common;
+using SignContract;
+
 internal sealed class ContractBuilder
 {
     public static ContractBuilder Create() => new();
 
-    private DateTimeOffset _preparedAt;
+    private DateTimeOffset? _preparedAt;
 
     public ContractBuilder PreparedAt(DateTimeOffset preparedAt)
     {
@@ -12,16 +15,17 @@ internal sealed class ContractBuilder
         return this;
     }
 
-    public PreparedContractBuilder Prepared() => new(Build());
+    public SignContractBuilder Sign() => new(Build());
 
     private Contract Build()
     {
+        var preparedAt = _preparedAt ?? FakeContractDates.PreparedAt;
         var prepareContractParameters = PrepareContractParameters.GetValid();
         var contract = Contract.Prepare(
             Guid.NewGuid(),
             prepareContractParameters.MaxAge,
             prepareContractParameters.MaxHeight,
-            _preparedAt);
+            preparedAt);
 
         return contract;
     }
