+++ Chapter-4-applying-tactical-domain-driven-design/Fitnet.Contracts/Src/Fitnet.Contracts.Core.UnitTests/TerminateBindingContract/TerminateBindingContractTests.cs
@@ -14,8 +14,7 @@ public sealed class TerminateBindingContractTests
     {
         BindingContract bindingContract = ContractBuilder
             .Create()
-                .PreparedAt(FakeContractDates.PreparedAt)
-            .Prepared()
+            .Sign()
                 .SignedOn(FakeContractDates.SignDay, _fakeToday);
 
         bindingContract.Terminate(_terminatedAt);
